Removing WebPushFrameReader and replace it with a listener.
